chore: move disconnect action to its own file
