+++ src/actions/disconnect.ts
@@ -1,0 +1,133 @@
+/**
+ * @license
+ * Copyright 2025 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import {
+  ASTNode,
+  RenderedConnection,
+  ShortcutRegistry,
+  utils as BlocklyUtils,
+} from 'blockly';
+import * as Constants from '../constants';
+import type {WorkspaceSvg} from 'blockly';
+import {Navigation} from '../navigation';
+
+const KeyCodes = BlocklyUtils.KeyCodes;
+
+/**
+ * Action to insert a block into the workspace.
+ *
+ * This action registers itself as both a keyboard shortcut and a context menu
+ * item.
+ */
+export class DisconnectAction {
+  /**
+   * Function provided by the navigation controller to say whether editing
+   * is allowed.
+   */
+  private canCurrentlyEdit: (ws: WorkspaceSvg) => boolean;
+
+  /**
+   * Registration name for the keyboard shortcut.
+   */
+  private shortcutName = Constants.SHORTCUT_NAMES.DISCONNECT;
+
+  constructor(
+    private navigation: Navigation,
+    canEdit: (ws: WorkspaceSvg) => boolean,
+  ) {
+    this.canCurrentlyEdit = canEdit;
+  }
+
+  /**
+   * Install this action as both a keyboard shortcut and a context menu item.
+   */
+  install() {
+    this.registerShortcut();
+  }
+
+  /**
+   * Uninstall this action as both a keyboard shortcut and a context menu item.
+   * Reinstall the original context menu action if possible.
+   */
+  uninstall() {
+    ShortcutRegistry.registry.unregister(this.shortcutName);
+  }
+
+  /**
+   * Create and register the keyboard shortcut for this action.
+   */
+  private registerShortcut() {
+    const disconnectShortcut: ShortcutRegistry.KeyboardShortcut = {
+      name: this.shortcutName,
+      preconditionFn: (workspace) => this.canCurrentlyEdit(workspace),
+      callback: (workspace) => {
+        switch (this.navigation.getState(workspace)) {
+          case Constants.STATE.WORKSPACE:
+            this.disconnectBlocks(workspace);
+            return true;
+          default:
+            return false;
+        }
+      },
+      keyCodes: [KeyCodes.X],
+    };
+    ShortcutRegistry.registry.register(disconnectShortcut);
+  }
+
+  /**
+   * Disconnects the connection that the cursor is pointing to, and bump blocks.
+   * This is a no-op if the connection cannot be broken or if the cursor is not
+   * pointing to a connection.
+   *
+   * @param workspace The workspace.
+   */
+  disconnectBlocks(workspace: WorkspaceSvg) {
+    const cursor = workspace.getCursor();
+    if (!cursor) {
+      return;
+    }
+    let curNode: ASTNode | null = cursor.getCurNode();
+    let wasVisitingConnection = true;
+    while (curNode && !curNode.isConnection()) {
+      curNode = curNode.out();
+      wasVisitingConnection = false;
+    }
+    if (!curNode) {
+      console.log('Unable to find a connection to disconnect');
+      return;
+    }
+    const curConnection = curNode.getLocation() as RenderedConnection;
+    if (!curConnection.isConnected()) {
+      return;
+    }
+    const superiorConnection = curConnection.isSuperior()
+      ? curConnection
+      : curConnection.targetConnection!;
+
+    const inferiorConnection = curConnection.isSuperior()
+      ? curConnection.targetConnection!
+      : curConnection;
+
+    if (inferiorConnection.getSourceBlock().isShadow()) {
+      return;
+    }
+
+    if (!inferiorConnection.getSourceBlock().isMovable()) {
+      return;
+    }
+
+    superiorConnection.disconnect();
+    inferiorConnection.bumpAwayFrom(superiorConnection);
+
+    const rootBlock = superiorConnection.getSourceBlock().getRootBlock();
+    rootBlock.bringToFront();
+
+    if (wasVisitingConnection) {
+      const connectionNode = ASTNode.createConnectionNode(superiorConnection);
+      workspace.getCursor()!.setCurNode(connectionNode!);
+    }
+  }
+}
